+++ cypress.config.ts
@@ -1,6 +1,8 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { defineConfig } from 'cypress';
 
+import initConfig from './cypress/plugins';
+
 export default defineConfig({
   video: false,
   retries: {
@@ -11,8 +13,7 @@ export default defineConfig({
     // We've imported your old cypress plugins here.
     // You may want to clean this up later by importing these.
     setupNodeEvents(on, config) {
-      // eslint-disable-next-line global-require, @typescript-eslint/no-var-requires, import/extensions
-      return require('./cypress/plugins/index.ts')(on, config);
+      return initConfig(on, config);
     },
     baseUrl: 'http://localhost:3111',
   },
