+++ src/components/App.tsx
@@ -5,8 +5,8 @@ import { saveUrlForRedirection } from '@graasp/sdk';
 import { CustomInitialLoader, withAuthorization } from '@graasp/ui';
 
 import { DOMAIN } from '@/config/env';
+import { SIGN_IN_PATH } from '@/config/externalPaths';
 
-import { SIGN_IN_PATH } from '../config/constants';
 import {
   FAVORITE_ITEMS_PATH,
   HOME_PATH,
